requirements.txt
asgiref==3.8.1
certifi==2025.1.31
Django==5.2
mysqlclient==2.2.7
Pillow==10.0.0
django-admin-interface==0.30.0
django-ckeditor==6.7.0
django-markdownx>=4.0.7,<5
Markdown>=3.5
django-colorfield==0.13.0
django-crispy-forms==2.0
langchain==0.2.1
langchain-openai==0.1.7
python-dotenv==1.0.1
SQLAlchemy==2.0.40
tqdm==4.67.1
